Adjusted list_ideas_spec let block, divided PageModels module classes into separate files
